feat: use docker events for faster test feedback (#136)
* feat: use docker events for better reliability.

Adds a docker event bus that can be used for improving the dockest experience. It is currently used in the following ways: Wait for container start before trying to resolve the container id (which solves possible timeouts due to first downloading the image) and stop resolving the container id in case the container unexpectedly dies. This results in less flakiness. In the future the readiness check API could be based upon docker events (see https://github.com/erikengervall/dockest/issues/85#issuecomment-541052423).

* update snapshots

* implicitly return undefined right away

* resolve container id via docker events

* abort check connection in case the container dies

* abort healthcheck in case container unexpectedly dies.

* add healthcheck for checking whether the health_status event is emitted.

* apply review feedback

* add more tests

Co-authored-by: Erik Engervall <erik.engervall@gmail.com>